create_sorter_utils
